extras/wificfg: make the default wifi options overridable
wificfg.c defines the default values for ssid, password and hostname. But
because these are not weak, trying to redefine these causes a link error.

This change marks the definitions in wificfg.c weak, thus making it
possible for a program using the wificfg component to redefine these w/o
having to touch wificfg.c.